FIX: UNIX socket stop listening when worker number decreases
The UNIX socket is only listened by the last forked worker.
when worker number decreases, that (old) last worker closes
the listening UNIX socket. And the newly spawned workers inherit
corresponding socket FD, but do NOT unlink old unix sock file and
re-bind the socket.

In short, no worker is listening on the unix sockets after reload.
We need to make sure only one worker is binding and listening
the unix socket. Here we choose the worker of `id` 0 to be the one.